fix(job): insert try catch

+++ src/services/job.service.js
@@ -78,8 +78,13 @@ async function deleteJobByJobId(req, res, next) {
   try {
     const { jobId } = req.body;
 
-    const deleteJob = await Job.findOneAndDelete({ _id: jobId });
-    res.status(200).json({ message: 'Job deleted successfully', data: deleteJob });
+    const job = await Job.findOne({ _id: jobId });
+    if (job) {
+      const deleteJob = await Job.findOneAndDelete({ _id: jobId });
+      res.status(200).json({ message: 'Job deleted successfully', data: deleteJob });
+    } else {
+      res.status(404).json({ message: 'Job not found' });
+    }
 
     await session.commitTransaction();
     session.endSession();
